+++ towerdiscord/integration_test.go
@@ -2,6 +2,7 @@ package towerdiscord_test
 
 import (
 	"context"
+	"fmt"
 	"github.com/tigorlazuardi/tower"
 	"github.com/tigorlazuardi/tower/bucket"
 	"github.com/tigorlazuardi/tower/internal/loader"
@@ -15,8 +16,9 @@ import (
 var _ towerdiscord.Hook = (*testHook)(nil)
 
 type testHook struct {
-	t  *testing.T
-	wg *sync.WaitGroup
+	t                  *testing.T
+	wg                 *sync.WaitGroup
+	checkBucketContext bool
 }
 
 func (t testHook) PreMessageHook(ctx context.Context, _ *towerdiscord.WebhookContext) context.Context {
@@ -35,6 +37,15 @@ func (t testHook) PostMessageHook(ctx context.Context, _ *towerdiscord.WebhookCo
 		}
 	} else {
 		t.t.Error("context value of test should exist in PostMessageHook")
+	}
+	if t.checkBucketContext {
+		if e, ok := ctx.Value("test-bucket").(string); ok {
+			if e != "test-bucket" {
+				t.t.Errorf("context value of test should have value of 'test-bucket' not '%s'", e)
+			}
+		} else {
+			t.t.Error("context value of test-bucket should exist in PostMessageHook")
+		}
 	}
 }
 
@@ -62,7 +73,7 @@ func (t testHook) PostBucketUploadHook(ctx context.Context, _ *towerdiscord.Webh
 			t.t.Errorf("context value of test should have value of 'test' not '%s'", e)
 		}
 	} else {
-		t.t.Error("context value of test should exist in PreBucketUploadHook")
+		t.t.Error("context value of test should exist in PostBucketUploadHook")
 	}
 	if e, ok := ctx.Value("test-bucket").(string); ok {
 		if e != "test-bucket" {
@@ -104,13 +115,20 @@ func TestIntegration(t *testing.T) {
 	bot.SetName("tower-discord-integration-test")
 	bot.SetHook(testHook{t: t, wg: wg})
 	tow.RegisterMessenger(bot)
-	tow.NewEntry("test %d", 123).Context(tower.F{"foo": "bar", "struct": foo{}}).Notify(ctx)
+	//tow.NewEntry("test %d", 123).Context(tower.F{"foo": "bar", "struct": foo{}}).Notify(ctx)
 	origin := tow.Wrap(foo{FooMessage: "something > something < something & Bad Request"}).Code(400).Message("this is bad request error").Context(tower.F{
 		"light": map[string]any{"year": 2021, "month": "january"},
 		"bar":   "baz",
 	}).Freeze()
 	wrapped := tow.WrapFreeze(origin, "wrapping error")
 	_ = tow.Wrap(wrapped).Message("wrapping error").Context(tower.F{"wrapping": 123, "nil_value": nil}).Notify(ctx)
+
+	const loremIpsum = "lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.\n\n"
+	fields := tower.F{}
+	for i, j := 0, 0; i < 4096; i, j = i+len(loremIpsum), j+1 {
+		fields[fmt.Sprintf("field_%d", j)] = loremIpsum
+	}
+	tow.NewEntry("test big text").Context(fields).Notify(ctx)
 	err := bot.Wait(ctx)
 	if err != nil {
 		t.Error(err)